langsetry_webapp/requirements.txt
flask
google-genai
python-dotenv
markdown2 